refactor : string관련 테스트 패키지 변경
