+++ src/test/java/study/string/SubStringTest.java
@@ -1,4 +1,4 @@
-package study;
+package study.string;
 
 import static org.assertj.core.api.Assertions.assertThat;
